Get prompt colors working again
There is a lot of code that executes on the condition that the zsh version is 4.0 or 4.2, even though the latest versions of zsh are 5.x. This needs to be cleaned up. For now, I'm removing the conditional execution controls for "autoload colors" so that I can have my prompt colors back again.